refactor(bayn): modularize market data
